# Possible alternative to our latin1_to_utf8_avx512 routine.