We only change the bulk of the processing, leaving the tail intact. It is currently no faster.

GCC 12, Icelake.

Main branch:

```
convert_latin1_to_utf8+icelake, input size: 199331, iterations: 30000, dataset: unicode_lipsum/wikipedia_mars/german.latin1.txt
   0.318 ins/byte,    0.136 cycle/byte,   22.917 GB/s (3.0 %),     3.111 GHz,    2.340 ins/cycle 
   0.318 ins/char,    0.136 cycle/char,   22.917 Gc/s (3.0 %)     1.00 byte/char 
convert_latin1_to_utf8+icelake, input size: 271743, iterations: 30000, dataset: unicode_lipsum/wikipedia_mars/portuguese.latin1.txt
   0.406 ins/byte,    0.160 cycle/byte,   19.369 GB/s (1.8 %),     3.104 GHz,    2.533 ins/cycle 
   0.406 ins/char,    0.160 cycle/char,   19.369 Gc/s (1.8 %)     1.00 byte/char 
```

This PR:

```
convert_latin1_to_utf8+icelake, input size: 199331, iterations: 30000, dataset: unicode_lipsum/wikipedia_mars/german.latin1.txt
   0.323 ins/byte,    0.137 cycle/byte,   22.651 GB/s (2.6 %),     3.112 GHz,    2.349 ins/cycle 
   0.323 ins/char,    0.137 cycle/char,   22.651 Gc/s (2.6 %)     1.00 byte/char 
convert_latin1_to_utf8+icelake, input size: 271743, iterations: 30000, dataset: unicode_lipsum/wikipedia_mars/portuguese.latin1.txt
   0.414 ins/byte,    0.165 cycle/byte,   18.838 GB/s (1.4 %),     3.104 GHz,    2.514 ins/cycle 
   0.414 ins/char,    0.165 cycle/char,   18.838 Gc/s (1.4 %)     1.00 byte/char 
```